refactor: improve filters
