+++ apps/api/src/orpc/routers/proxy/index.ts
@@ -1,5 +1,5 @@
 import type { DatabaseQueryResult } from '@conar/shared/databases'
-import type { QueryParams } from '@conar/shared/utils/sql'
+import type { QueryParams } from '@conar/shared/filters/sql'
 import process from 'node:process'
 import { parseConnectionString } from '@conar/connection'
 import { readSSLFiles } from '@conar/connection/server'
