+++ apps/desktop/src/entities/database/sql/total.ts
@@ -1,4 +1,4 @@
-import type { ActiveFilter } from '@conar/shared/utils/filters'
+import type { ActiveFilter } from '@conar/shared/filters'
 import type { databases } from '~/drizzle'
 import { type } from 'arktype'
 import { count, sql } from 'drizzle-orm'
