+++ apps/desktop/src/routes/(protected)/_protected/database/$id/sql/-chat.ts
@@ -3,7 +3,7 @@ import type { InferToolInput, InferToolOutput } from 'ai'
 import type { chatsMessages, databases } from '~/drizzle'
 import { Chat } from '@ai-sdk/react'
 import { convertToAppUIMessage } from '@conar/shared/ai-tools'
-import { SQL_FILTERS_LIST } from '@conar/shared/utils/sql'
+import { SQL_FILTERS_LIST } from '@conar/shared/filters/sql'
 import { eventIteratorToStream } from '@orpc/client'
 import { lastAssistantMessageIsCompleteWithToolCalls } from 'ai'
 import { v7 as uuid } from 'uuid'
