+++ apps/desktop/src/routes/(protected)/_protected/database/$id/table/-components/header-search.tsx
@@ -1,5 +1,5 @@
-import type { ActiveFilter } from '@conar/shared/utils/filters'
-import { SQL_FILTERS_LIST } from '@conar/shared/utils/sql'
+import type { ActiveFilter } from '@conar/shared/filters'
+import { SQL_FILTERS_LIST } from '@conar/shared/filters/sql'
 import { Button } from '@conar/ui/components/button'
 import { ContentSwitch } from '@conar/ui/components/custom/content-switch'
 import { LoadingContent } from '@conar/ui/components/custom/loading-content'
