+++ apps/desktop/src/routes/(protected)/_protected/database/$id/table/index.tsx
@@ -1,7 +1,7 @@
-import type { ActiveFilter } from '@conar/shared/utils/filters'
+import type { ActiveFilter } from '@conar/shared/filters'
 import type { Store } from '@tanstack/react-store'
 import type { storeState } from './-store'
-import { SQL_FILTERS_GROUPED } from '@conar/shared/utils/sql'
+import { SQL_FILTERS_GROUPED } from '@conar/shared/filters/sql'
 import { title } from '@conar/shared/utils/title'
 import { ResizableHandle, ResizablePanel, ResizablePanelGroup } from '@conar/ui/components/resizable'
 import { createFileRoute } from '@tanstack/react-router'
