+++ packages/shared/package.json
@@ -7,9 +7,10 @@
     "./constants": "./src/constants.ts",
     "./encryption": "./src/encryption.ts",
     "./enums/*": "./src/enums/*.ts",
+    "./filters": "./src/filters/index.ts",
+    "./filters/*": "./src/filters/*.ts",
     "./databases": "./src/databases.ts",
     "./schemas/postgres/*": "./src/schemas/postgres/*.ts",
-    "./sql/*": "./src/sql/*.ts",
     "./utils/*": "./src/utils/*.ts"
   },
   "scripts": {
